% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateScenerioOutputs.R
\name{aggregateScenerioOutputs}
\alias{aggregateScenerioOutputs}
\title{aggregateScenerioOutputs
This function takes outputs from 'importAndCondense' and aggregates results by site, resulting in a final data set which has proportion of time spent in each state, proportion of transmissions for each site, along with information on the site category.
Possible mechanisms for transmissions are:
\enumerate{
\item LFM = live fish movement
\item RB = river-based transmission
\item SDM = short-distance mechanical
\item DIM = distance-independent mechanical (random spillover)
}}
\usage{
aggregateScenerioOutputs(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) name of the scenario assigned in the \code{params.yaml}.}
}
\value{
(class data.table) data.table containing time spend consecutively in each of the
simulation states for each site and the proportion for each tranmission route per site.
This contains the following information:
\enumerate{
\item \code{site_id} (integer) the site identification number.
\item \code{t_infected_S1} (numeric) total time (days) in which the site has been in state 1
(infected) through all of the simulations
\item \code{proportion_state1} (numeric) proportion of time (days) in which the site has been
in state 1 (infected) throughout the course of all the simulations
\item \code{t_infected_S3} (numeric) total time (days) in which the site has been in state 3
(Infected with secondary controls) through all of the simulations
\item \code{proportion_state3} (numeric) proportion of time (days) in which the site has been
in state 3 (Infected with secondary controls) throughout the course of all the simulations
\item \code{t_infected_S0} (numeric) total time (days) in which the site has been in state 0 (uninfected)
through all of the simulations
\item \code{proportion_state0} (numeric) proportion of time (days) in which the site has been in state 0
(Uninfected) throughout the course of all the simulations
\item \code{t_infected_S2} (numeric) total time (days) in which the site has been in state 2 (Uninfected
with secondary controls) through all of the simulations
\item \code{proportion_state2} (numeric) proportion of time (days) in which the site has been in state 2
(Uninfected with secondary controls) throughout the course of all the simulations
\item \code{LFM}(numeric) number of transmissions due to LFMs for each site across all simulations
\item \code{RB}(numeric) number of transmissions due to RBs for each site across all simulations
10.\code{SDM}(numeric) number of transmissions due to SDM for each site across all simulations
11.\code{DIM}(numeric) number of transmissions due to DIM for each site across all simulations
12.\code{siteTotalTransmissions}(numeric) total number of transmissions for each site across all simulations
13.\code{LFM_prop}(numeric) proportion of transmissions due to LFMs for each site across all simulations
14.\code{RB_prop} (numeric) proportions of transmissions due to RB for each site across all simulations
15.\code{SDM_prop}(numeric) proportions of transmissions due to SDM for each site across all simulations
16.\code{DIM_prop}(numeric) proportions of transmissions due to DIM for each site across all simulations
}
}
\description{
Possible states are:
0 = Uninfected
1 = Infected
2 = Uninfected with secondary controls
3 = Infected with secondary controls
}
