% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRandomSpillover.R
\name{calcRandomSpillover}
\alias{calcRandomSpillover}
\title{calcRandomSpillover}
\usage{
calcRandomSpillover(
  clinical_state_vector,
  spread_restricted_off,
  spread_restricted_on,
  site_indices,
  trans_name,
  run_time_params,
  trans_type
)
}
\arguments{
\item{clinical_state_vector}{(class numeric) numeric binary vector of length number of sites
containing information on whether each site is in an 1 = clinically infected or 0 =
non-clinically infected state (the latter includes both susceptible and latently infected sites).
Note: vector created within the \code{update_rate} function.}

\item{spread_restricted_off}{(class logical) logical vector of length number of sites that
states whether movements off this site are currently restricted (TRUE) or unrestricted (FALSE).
(Note: created in the \code{update_rate} function of aquanet-mod).}

\item{spread_restricted_on}{(class logical) logical vector of length number of sites that
states whether movements on to this site are currently restricted (TRUE) or unrestricted (FALSE).
(Note: created in the \code{update_rate} function of aquanet-mod).}

\item{site_indices}{(class numeric) vector of 0-based site indices of length 'number of sites'.}

\item{trans_name}{(class string) string in quotation marks "" stating transition type. Note: this
should correspond to a column name in the input parameter file and is case sensitive. An example
input may be "Fomite_Transmission_Independent_Prob".}

\item{run_time_params}{(class data frame) of model run time parameters imported from original
parameter file which is subsequently split into model set up and model run time parameters. Data
frame contains probabilities from the scenarios of interest for listing transition rates.}

\item{trans_type}{(class numeric) single numeric value indicating the type of transition occurring.}
}
\value{
(class list) of either length 4 or 5 containing:
\enumerate{
\item (class numeric) numeric vector of transition types (\code{trans_type}) with length 'number of sites
that are infected with no restrictions on movement off site'.
\item (class integer) integer vector of 0-based indices of site subject to transition i.e.
susceptible (positions/coordinate x of site randomly sampled from susceptible sites with no
restrictions to prevent spread on site) with length 'number of sites that are infected with no
restrictions on movement off site'.
\item (class numeric) numeric vector of transition rate/transmission probability (the probability
of risk contact at sites connected to susceptible sites) calculated by
\code{1 / run_time_params[[trans_name]]} with length 'number of sites that are infected with no
resrictions on movement off site'.
\item (class logical) logical vector containing NA's to depict unknown infection source with length
'number of sites that are infected with no movement off site'.
\item (class integer) number of sites that are infected with no restrictions on movements off site.
}
}
\description{
Generate a list of transition rate objects using an input vector of clinically infected sites
(\code{clinical_state_vector}) to extract sites that are susecptible to infection that do not have
restricted movements on site. Additionally, identify sites with spreading potential that are
clinically infected and have no restrictions on movements off site.
}
\details{
Create list of vectors with length equivalent to the number of sites with spreading potential and
containing: (1) transition type vector depicted by the column number of \code{trans_name} within the
\code{run_time_params}, (2) randomly selected transitioning/susceptible site vector, (3) transition
rate/probability vector calculated using \code{1 / run_time_params[[trans_name]]}, and (4) source site
vector of NAs. This list is populated with NULL if there are no susceptible sites. If there are
1 or more susceptible sites, the fifth element of this list is the number of infected sites.

Note: in aquanet-mod, this function is called multiple times exclusively within the \code{update_rate}
function.

TODO: update function names if changed
}
