% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverTransmission.R
\name{calcRiverTransmission}
\alias{calcRiverTransmission}
\title{calcRiverTransmission}
\usage{
calcRiverTransmission(
  matrix_river_distances_prob,
  clinical_state_vector,
  move_restricted_off,
  move_restricted_on,
  trans_type
)
}
\arguments{
\item{matrix_river_distances_prob}{(class dgTmMatrix, Matrix package) sparse matrix containing
probability of transmission between sites connected via the river network by river water. Note:
this is the second element in the list output of
\code{aquanet::createRiverDistanceProbabilityMatrix()}.}

\item{clinical_state_vector}{(class numeric) numeric binary vector of length number of sites
containing information on whether each site is in an 1 = clinically infected or 0 =
non-clinically infected state (the latter includes both susceptible and latently infected sites).
Note: vector created within the \code{update_rate} function.}

\item{move_restricted_off}{(class logical) logical vector of length number of sites that
states whether movements off this site are currently restricted (TRUE) or unrestricted (FALSE).
(Note: created in the \code{update_rate} function of aquanet-mod).}

\item{move_restricted_on}{(class logical) logical vector of length number of sites that
states whether movements on to this site are currently restricted (TRUE) or unrestricted (FALSE).
(Note: created in the \code{update_rate} function of aquanet-mod).}

\item{trans_type}{(class numeric) number stating transition type.}
}
\value{
(class list) of length 5 from \code{aquanet::listRatesSusceptibleRiskContacts} containing:
\enumerate{
\item (class numeric) numeric vector of transition types (input \code{trans_type}) with length 'number of
sites with contact probability greater than 0 where the receiving site is in a susceptible state'.
\item (class integer) integer vector of sites subject to transition (positions/coordinate x of sites
with risk contact at susceptible sites \code{(spmatrix_risk_contacts@i)}) with length 'number of
sites with contact probability greater than 0 where the receiving site is in a susceptible state'.
\item (class numeric) numeric vector of transition rates/transmission probability (the probability
of risk contact at sites connected to susceptible sites. \code{(spmatrix_risk_contacts@x)}) with
length 'number of sites with contact probability greater than 0 where the receiving site is in a
susceptible state'.
\item (class integer) numeric vector of source sites (positions/coordinate y of sites with risk
contact at susceptible sites \code{(spmatrix_risk_contacts@j)}) with length 'number of sites with
contact probability greater than 0 where the receiving site is in a susceptible state'.
\item (class integer) number of sites that have a risk contact probability above 0 where the
receiving site is in a susceptible state.
}
}
\description{
Take the river transmission probability matrix and correct source to receiving site transmission
probabilities so only connections where the source site is clinically infected
(\code{state_vector = 1}) and where movements off site are NOT prevented (i.e. movements off site are
allowed) have a transmission probability greater than 0. In addition. correct receiving from
source site transmission probabilities so only connections where the receiving site has NO
restrictions on on site movements.
}
\details{
Use the corrected river transmission probabilities matrix to extract a list of information
related to the sites that have the potential to be infected using
\code{aquanet::listRatesSusceptibleRiskContacts()}.

TODO: update function names if changed (listRatesSusceptibleRiskContacts/listInfectionRates)
}
