% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorisedSites.R
\name{categorisedSites}
\alias{categorisedSites}
\title{categorisedSites}
\usage{
categorisedSites(lfmData, ProductionData)
}
\arguments{
\item{lfmData}{(data.table) Live Fish Movement data extracted from STARFISH database with columns 'Src_Code', 'Scr_AutRegStatus', 'PurposeOfMovement' included.}

\item{ProductionData}{(data.table) Production data as data.table (using function 'fread(dataset)') extracted from STARFISH, with columns 'Code', 'AuthorisationAndRegistrationStatusCode', 'CodeDescription2' included.}
}
\value{
a csv file with site categorised by type; 'Farm/fishery', 'Table', 'Restocker', 'Ongrower', 'Hatchery', 'Fishery'.
A site can fall into multiple categories.
}
\description{
Sites from fisheries and farms are labeled if they fall into the following site type; Table, Restocker, Ongrower, Hatchery and Fishery.
Using both the LFM and Production datasets. A csv file is then outputed with the classifications for each site labelled using 'Y' and 'N'
based on if they meet the conditions for each site type.
}
