% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCatchmentLevelRestocking.R
\name{checkCatchmentLevelRestocking}
\alias{checkCatchmentLevelRestocking}
\title{checkCatchmentLevelRestocking}
\usage{
checkCatchmentLevelRestocking(
  control_matrix,
  spmatrix_sites_catchment,
  n_catchments
)
}
\arguments{
\item{control_matrix}{(class matrix) matrix containing 7 columns depicting different control
states and rows (of length number of sites) depicting whether each sites is 1 = in the specified
control state or 0 = not in the specified control state.}

\item{spmatrix_sites_catchment}{(class dgCMatrix, Matrix package) sparse matrix containing site
to catchment summary.
Note: this is created with the \code{aquanet::createCatchmentToSiteMatrix()} function.}

\item{n_catchments}{(class integer) number of catchments.}
}
\value{
(class list) of length 3 containing:
\enumerate{
\item (class matrix) input \code{control_matrix} containing 7 columns depicting different control
states and rows (of length number of sites) depicting whether each sites is 1 = in the specified
control state or 0 = not in the specified control state. This matrix is unedited and output to
keep a record of the input data used to generate these catchment restocking data.
\item (class logical) vector of length number of catchments (\code{n_catchments}) depicting whether
catchments contain fallow sites.
\item (class logical) vector of length number of catchments (\code{n_catchments}) depicting whether
catchments contain only post-fallow sites ready for restocking with no fallow sites.
}
}
\description{
This function identifies catchments that are ready to restock as all sites within them are in a
post-fallow state (see details).
}
\details{
TODO: control matrix not altered here - remove output?
TODO: is catchments_with_fallow_some output used?

Function takes control matrix stating which sites are in a fallow state and which are in a post-
fallow state (ready for restocking). This information is combined with the siteID-to-catchment
matrix to calculate the number of sites in a fallow or post-fallow state within each catchment.
This information is then used to determine which catchments contain only sites in a post-fallow
state that are ready for restocking and to determine which catchments contain sites in a fallow
state.
}
