% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTransitionRates.R
\name{combineTransitionRates}
\alias{combineTransitionRates}
\title{combineTransitionRates}
\usage{
combineTransitionRates(list_append, list_base)
}
\arguments{
\item{list_append}{(class list) of length at least 4 containing the following at list positions
1, 2, 3, and 4:
\enumerate{
\item (class numeric) vector of transition types.
\item (class integer) vector of sites subject to transition.
\item (class numeric) vector of transition rates (transmission probability).
\item (class integer) vector of source sites (of disease in case of transmission).
contact.
}}

\item{list_base}{(class list) of length 4 with either empty or populated list positions
containing the following at list positions 1, 2, 3, and 4:
\enumerate{
\item (class numeric) vector of transition types.
\item (class integer) vector of sites subject to transition.
\item (class integer) vector of transition rates (transmission probability).
\item (class numeric) vector of source sites (of disease in case of transmission).
}}
}
\value{
(class list) of length 4 containing the following concatenated information from
\code{list_append} and \code{list_base}:
\enumerate{
\item (class numeric) vector of transition types.
\item (class integer) vector of sites subject to transition.
\item (class integer) vector of transition rates (transmission probability).
\item (class numeric) vector of source sites (of disease in case of transmission).
}
}
\description{
Append the contents of \code{list_append} list elements to the corresponding list elements in
\code{list_base}. Note, that list elements should be in the order specified within the inputs.
}
\details{
Note: in aquanet-mod, this function is called multiple times exclusively within the
\code{aquanet::updateRates} function.
}
