% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commitResults.R
\name{commitResults}
\alias{commitResults}
\title{commitResults}
\usage{
commitResults(
  df_states,
  n_states,
  n_sites,
  site_indices,
  commit_int,
  iteration_vector,
  batch_num,
  simulation_num,
  save_num,
  filepath_results
)
}
\arguments{
\item{df_states}{(class data.table data.frame) data frame of zeros with dimensions number of rows
= number of sites (\code{n_sites}) + number of states (\code{n_states}), and number of columns = commit
interval (\code{commit_int}) + 1.}

\item{n_states}{(class numeric) number of different combinations of states possible within the
model.}

\item{n_sites}{(class numeric) number of sites within the contact network (model run).}

\item{site_indices}{(class integer) vector of 0-based site indices of length number of sites
(\code{n_sites}).}

\item{commit_int}{(class numeric) number of intervals at which results should be committed/saved.}

\item{iteration_vector}{(class integer) integer vector of length \code{commit_int} containing values
1:commit interval (\code{commit_int}).}

\item{batch_num}{(class numeric) batch number for model run.}

\item{simulation_num}{(class numeric) simulation number for model run.}

\item{save_num}{(class numeric) number of full saves for model run.}

\item{filepath_results}{(class string) path to results directory for model run.}
}
\value{
Saved .RData file containing two data frames located within the \code{filepath_results}
Full_Details results directory. Data frame 1: \code{sim_states} contains site ID and infection and
control status for every time step within a simulation.  Data frame 2: \code{sim_times} contains the
time step ID, simulation number and details of time step size.
}
\description{
Output information on sites that become infected or subject to control measures at each time step
within each simulation of the model. Additionally, output information on size of each time step
within the model simulations. Results are saved as an .RData file within the Full_Details
directory.
}
\details{
TODO: change Full_Details path to match Sarah's folder structure
}
