% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCatchmentLookupTables.R
\name{createCatchmentToSiteMatrix}
\alias{createCatchmentToSiteMatrix}
\title{createCatchmentToSiteMatrix}
\usage{
createCatchmentToSiteMatrix(graph, filename_catchment_layer)
}
\arguments{
\item{graph}{(class igraph) Graph of connections/movements between sites produced with iGraph (using script importSiteData.R of AquaNet-Mod). This includes both live fish movements and Section 30 movements.}

\item{filename_catchment_layer}{(class string) String containing the file path and file name for .shp file containing catchment information.}
}
\value{
(class list) of length 2 containing (1) data frame of site to catchment information and (2) dgCMatrix sparse matrix containing site to catchment summary.
}
\description{
Extract connectivity matrix (graph) information to produce a data frame of siteID linked to catchment code (TRUNK_CODE) and merge with detailed catchment information from GIS layer data (.shp). This produced list output 1 which which maps site to catchment. To produce list output 2, convert the catchment (columns) to site (rows) information to a sparse matrix. Note: ensure that the order of sites matches that included within the contact matrix.
}
