% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContactProbabilityMatrixTopSitesRemoved.R
\name{createContactProbabilityMatrixTopSitesRemoved}
\alias{createContactProbabilityMatrixTopSitesRemoved}
\title{createContactProbabilityMatrixTopSitesRemoved}
\usage{
createContactProbabilityMatrixTopSitesRemoved(graph, movement_period, n_remove)
}
\arguments{
\item{graph}{(class igraph) Graph of connections/movements between sites produced with iGraph
(using script importSiteData.R of AquaNet-Mod). This includes both live fish movements and
Section 30 movements.}

\item{movement_period}{(class numeric) The period of time (in days) for which movement data (live
fish movements and section 30 movements) were collected and used to create \code{graph} parameter.}

\item{n_remove}{(class numeric) The number of topmost connected sites to be removed from the contact
probability matrix.}
}
\value{
(class list) of length 3 containing (1) number of sites in movements matrix (integer),
(2) movements matrix (dgCMatrix, Matrix package), and (3) probability of movements matrix with
top sites zeroed (dgCMatrix, Matrix package).
}
\description{
Use the connectivity matrix (\code{graph}) to extract a movement matrix. Calculate the probability of
movement/contact between sites by dividing the number of movements by dividing by the period of
time (in days) for which movement information was used (\code{movement_period}). Extract sites with
greater than a user-specified quantile (\code{percentile}) of in (receiving) movements and out
(supplying) movements and overwrite their contact probabilities to zero to negate the impact of
top sites in the contact network. Then, determine the number of sites present in the contact
probability matrix. Return these three values as a list.
}
