% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRiverDistanceProbabilityMatrix.R
\name{createRiverDistanceProbabilityMatrix}
\alias{createRiverDistanceProbabilityMatrix}
\title{createRiverDistanceProbabilityMatrix}
\usage{
createRiverDistanceProbabilityMatrix(
  filepath_river_distances,
  out_createContactProbabilityMatrix
)
}
\arguments{
\item{filepath_river_distances}{(class string) String containing the file path and file name for
.csv containing information about distances between sites via river network proximity and
connectivity generated with GIS tool.}

\item{out_createContactProbabilityMatrix}{(class list) of length 3 containing (1) number of sites
in movements matrix (integer), (2) movements matrix (dgCMatrix, Matrix package), and (3)
probability of movements matrix with top sites zeroed (dgCMatrix, Matrix package).}
}
\value{
(class list) of length 2 containing (1) distances between sites on a river network (via
river connectivity) connections of 0 distances are removed (data frame) and (2) sparse matrix
containing probability of transmission between sites connected via the river network by river
water (dgTMatrix)
}
\description{
Generate a sparse matrix of site to site transmission probabilities via river distances between
sites connected by rivers (due to close proximity to the river network). River distances between
sites are loaded from .csv whose filename is specified as input \code{filepath_river_distances}.
Origin and destination site IDs are extracted, and the data subsetted to retain only connections
where both origin and destination site are present in the live fish movements contact matrix
(generated with igraph). Distances of zero are removed. Distances are converted to transmission
probabilities using the equation \code{0.005 * ((40 - (river_distance_length/1000))/39)} and
probabilities less than zero are corrected. A sparse matrix with site IDs in the same order as
the contact matrix is generated and populated with the river distance transmission probabilities.
The resulting list contains (1) the river distance data frame used and (2) a river distance
transmission probability sparse matrix.
}
