% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cullCost.R
\name{cullCost}
\alias{cullCost}
\title{cullCost}
\usage{
cullCost(farm_data, non_farm_data, cull_costs, site_types)
}
\arguments{
\item{farm_data}{(class data.frame) the \code{time_summary} data frame created from the
\code{timePerStage} function, for farms only.}

\item{non_farm_data}{(class data.frame) the \code{time_summary} data frame created from the
\code{timePerStage} function, for non-farms (fisheries) only.}

\item{cull_costs}{(class data.frame) of the costs to both the site and the
competent authority for each site type.}

\item{site_types}{(class list) a list of possible site types.}
}
\value{
(class data.frame) full_cull_cost_sim
}
\description{
Calculates a one-off cost to both the site and the competent authority in the
case of a site needing to be culled. Returns a data frame the
cost to cull across each simulation.
}
