% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeTimeSummary.R
\name{cumulativeTimeSummary}
\alias{cumulativeTimeSummary}
\title{cumulativeTimeSummary}
\usage{
cumulativeTimeSummary(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded.}
}
\value{
scenario_summary
}
\description{
Creates a summary of the cumulative number of sites infected by day for a given
scenario. Used to create time plots using the timePlots function.
}
