% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEvent.R
\name{doEvent}
\alias{doEvent}
\title{doEvent}
\usage{
doEvent(
  state_vector,
  control_matrix,
  transition_rates,
  tdiff,
  move_restricted_sites,
  non_peak_season,
  run_time_params,
  n_catchments,
  spmatrix_sites_catchment,
  time_vector,
  catchment_time_vector,
  catchments_with_post_fallow_only,
  source_inf_vector,
  source_inf_matrix,
  contact_tracing
)
}
\arguments{
\item{state_vector}{(class numeric) numeric binary vector of length number of sites containing
information about the state of each site in relation to a condition (E.g. is the site
1 = infected or 0 = susceptible state). This state vector should specify whether a site is in an
1 = infected or 0 = susceptible state. (Note: created within the \code{simulationCode} function for
loop).}

\item{control_matrix}{(class matrix) matrix containing 7 columns depicting different control
states and rows (of length number of sites) depicting whether each sites is 1 = in the specified
control state or 0 = not in the specified control state.}

\item{transition_rates}{(class list) of length 4 containing:
\enumerate{
\item (class numeric) vector of transition types.
\item (class integer) vector of sites subject to transition.
\item (class integer) vector of transition rates (transmission probability).
\item (class numeric) vector of source sites (of disease in case of transmission).
}}

\item{tdiff}{(class numeric) size of current time step generated in \code{simulationCode()} by picking
a random number from exponential distribution weighted by sum of transition rates.}

\item{move_restricted_sites}{(class logical) logical vector of length number of sites that
states whether movements at this site are currently restricted (TRUE) or unrestricted (FALSE).
(Note: created in the \code{update_rate} function of aquanet-mod and also input to
\code{excludeWithinCatchmentMovements} function).}

\item{non_peak_season}{(class logical) logical indicating whether the current timestep in the
model is within non peak season where transmission (of e.g. a pathogen) is lower.}

\item{run_time_params}{(class data frame) of model run time parameters imported from original
parameter file which is subsequently split into model set up and model run time parameters. Data
frame contains probabilities from the scenarios of interest for listing transition rates.}

\item{n_catchments}{(class integer) number of catchments.}

\item{spmatrix_sites_catchment}{(class dgCMatrix, Matrix package) sparse matrix containing site
to catchment summary.}

\item{time_vector}{(class numeric) vector of length 'number of sites' to record the amount of
simulation time that a site has remained in a susceptible/recovered or fallow state.}

\item{catchment_time_vector}{(class numeric) vector of length \code{n_catchments} to record the
amount of simulation time since every site in a catchment has been ready to be restocked.}

\item{catchments_with_post_fallow_only}{(class logical) vector of length number of catchments
(\code{n_catchments}) depicting whether catchments contain only post-fallow sites ready for restocking
with no fallow sites.}

\item{source_inf_vector}{(class numeric) vector of length 'number of sites' to track which sites
are responsible for infection when infection is transmitted via either Live Fish Movements or via
river network connectivity.}

\item{source_inf_matrix}{(class matrix) matrix of dimensions 'number of sites' x 'number of
sites' to track the sites that are infected either by Live Fish Movements of via river network
connectivity'. Note: this matrix is used for forward contact tracing.}

\item{contact_tracing}{(class logical) vector of length 1 indicating whether or not contact
tracing is taking place.}
}
\value{
(class list) of length 8 containing:
\enumerate{
\item (class numeric) \code{state_vector} numeric binary vector of length number of sites containing
information about the state of each site in relation to a condition (E.g. is the site
1 = infected or 0 = susceptible state). This state vector should specify whether a site is in an
1 = infected or 0 = susceptible state. (Note: created within the \code{simulationCode} function for
loop).
\item (class matrix) \code{control_matrix} updated matrix containing 7 columns depicting different
control states and rows (of length number of sites) depicting whether each sites is 1 = in the
specified control state or 0 = not in the specified control state.
\item (class numeric) \code{time_vector} updated vector of length 'number of sites' to record the amount
of simulation time that a site has remained in a susceptible/recovered or fallow state.
\item (class numeric) \code{catchment_time_vector} updated vector of length \code{n_catchments} to record the
amount of simulation time since every site in a catchment has been ready to be restocked.
\item (class logical) \code{catchments_with_post_fallow_only} vector of length number of catchments
(\code{n_catchments}) depicting whether catchments contain only post-fallow sites ready for restocking
with no fallow sites.
\item (class matrix) \code{source_inf_vector} updated matrix of dimensions 'number of sites' x 'number of
sites' to track the sites that are infected either by Live Fish Movements of via river network
connectivity'. Note: this matrix is used for forward contact tracing.
\item (class numeric) \code{rate_type} transition type for the site selected by \code{doEvent()} (extracted
from \code{transition_rates}).
\item (class matrix) \code{source_inf_matrix} matrix of dimensions 'number of sites' x 'number of
sites' to track the sites that are infected either by Live Fish Movements of via river network
connectivity'. Note: this matrix is used for forward contact tracing.
}
}
\description{
First, create variable a logical vector to store information on whether all sites in a catchment
have been in a post-fallow state for at least 4 days and are therefore ready to restock.
}
\details{
Next, extract required information from function inputs: sites that are fallow,  sites subject to
transition identified in \verb{transition rates[[2]]}, total number of possible site transitions, a
corrected probability of each transition occurring (number of transitions), and the number of
transition probabilities.

Identify sites that are under surveillance with no visible infection and for these sites and update
the \code{time_vector} by adding the time that has passed since the last step in the simulation
\code{tdiff}. This keeps record of the amount of time these sites are under surveillance.

To initiate the event, if there is more than one transition probability select a transition event
at random and for this event extract the site ID and rate type \code{trans_type} for the event.

Depending on the \code{trans_type} of the selected transition event update the infection status of the
site (whether it is listed as infected or susceptible in \code{state_vector}), update the controls in
place at the site, and if transmission occurs from a known source record this information. If
the site transitions from Controlled -> Fallow, or Infected -> Latent, or Latent -> Susceptible
via recovery or clearance then reset the time in time_vector.

Finally, update the time-dependent controls at the site within \code{control_matrix} for sites of
interest:

For sites that are in stage 1 surveillance and have not been reinfected check whether sites have
been under surveillance for required number days and transition those that have to stage 2
surveillance.

For sites that are in stage 2 surveillance and have not been reinfected check whether sites have
been under surveillance for required number of days and transition those that have to no disease
control state.

If sites have been in a fallow state for required number of days convert controls to post-fallow
state. Check catchment level restocking to identify catchments containing only post-fallow sites
and reset the \code{catchment_time_vector} to 0 for these. If these catchments have been in a ready
to restock state for greater than or equal to 4 days then move the sites within these catchments
to a no disease control state and reset the \code{time_vector} for these sites.

TODO: update function names
TODO: should 4 days be hard coded?
TODO: should update control 2 have !state_vector included in sites (i.e. not reinfected)
TODO: fix time_vector reset for controls and param definition
}
