% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeBetweennessSummary.R
\name{edgeBetweennessSummary}
\alias{edgeBetweennessSummary}
\title{Edge Betweenness summary
Summary of edge betweenness of contact network, which ranks which links are significant,
based on the fraction of shortest paths that pass through the links between network
nodes (susceptible contacts).}
\usage{
edgeBetweennessSummary(contact_network)
}
\arguments{
\item{contact_network}{(class igraph)  Graph of connections/movements between sites produced with iGraph
in '03_CreateContactNetwork.R' of AquaNet-mod.}
}
\value{
network_betweenness

network_betweenness
}
\description{
Edge Betweenness summary
Summary of edge betweenness of contact network, which ranks which links are significant,
based on the fraction of shortest paths that pass through the links between network
nodes (susceptible contacts).
}
