% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemicDuration.R
\name{epidemicDuration}
\alias{epidemicDuration}
\title{epidemicDuration}
\usage{
epidemicDuration(results)
}
\arguments{
\item{results}{(class data.frame) A data frame created from \code{loadResultsSummary}
Contains:
\enumerate{
\item \code{batch_no} the batch number
\item \code{k} the run number within the batch
\item \code{t} the model time
\item \code{t_diff} the difference between the current time and the time at the previous timestep
(the duration of the previous timestep)
\item \code{sim_no} the simulation number
\item \code{rate_type} the transition type
\item \code{no_controlled_catchments} the number of catchments under controls in that timestep
\item \code{cumulative_no_infected_sites} the cumulative number of infected sites in the simulation
Plus the number of sites in each state (1:39) plus redundancy (40:42)
}}
}
\value{
(class data.table) summary of epidemic duration (measured in days):
\enumerate{
\item \code{mean_duration} the mean epidemic duration for the scenario
\item \code{sd_duration} the standard deviation of the epidemic duration for the scenario
\item \code{min_duration} minimum epidemic duration
\item \code{max_duration} maximum epidemic duration
\item \code{median_duration} medium epidemic duration
\item \code{q05_duration} the 5\% quartile for epidemic duration
\item \code{q95_duration} the 95\% quartile for epidemic duration
\item \code{percent_time_out} the percentage of simulations for which the epidemic did not
last longer than 5 years (1800 model days)
\item \code{no_iter} numer of iterations/simulations run
}
}
\description{
epidemicDuration
}
