% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAndCondense.R
\name{importAndCondense}
\alias{importAndCondense}
\title{importAndCondense}
\usage{
importAndCondense(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) name of the scenario assigned in the \code{params.yaml}.}
}
\value{
(class data.table) data.table containing time spend consecutively in each of the
simulation states for each site and each simulation run. Multiple occurrences of a site in the
same state are counted separately. This contains the following information:
\enumerate{
\item \code{site_id} (integer) the site identification number.
\item \code{modelID} (integer) site idenfication number generated in model
\item \code{state} (integer) the state the site is in. TODO: add file with state definitions and link here
\item \code{group} (integer)
\item \code{sim_no} (integer) simulation number.
\item \code{timeID} (integer) simulation time step ID.
\item \code{t_total} (numeric) sum of time spend in state.
\item \code{t} (numeric) time at which change of state occurs.
\item \code{trans_type} (numeric) transition type that results in change of state.
\item onwards binary vectors indicating site type assigned in \code{site_types.csv}.
\item \code{row_sums} (integer) the sum of the site type vector, and is used to diagnose uncategorised
sites (usually fisheries).
}
}
\description{
Loads and condenses multiple the full data (full_details) output from a scenario run. Data are
grouped by simulation number and siteID. The amount of time each site spends in each state is
summed at each state change in the simulation.
}
