% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importsExports.R
\name{importsExports}
\alias{importsExports}
\title{Imports and exports
Speficy if connection between nodes is imports or exports-imports represent in-degree connections (the number of incoming connections), outputs represents the number of out-degree connections (the number of outgoing connections ). Then calculate summary stats of the movements}
\usage{
importsExports(contact_network, imports_exports)
}
\arguments{
\item{contact_network}{network simulates and tracks disease transmission from infected to susceptible contacts}

\item{imports_exports}{string input- either 'imports' or 'exports'}
}
\value{
movements
}
\description{
Imports and exports
Speficy if connection between nodes is imports or exports-imports represent in-degree connections (the number of incoming connections), outputs represents the number of out-degree connections (the number of outgoing connections ). Then calculate summary stats of the movements
}
