% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTransitionRates.R
\name{listTransitionRates}
\alias{listTransitionRates}
\title{listTransitionRates}
\usage{
listTransitionRates(
  run_time_params,
  state_vector,
  trans_name,
  site_indices,
  trans_type
)
}
\arguments{
\item{run_time_params}{(class data frame) of model run time parameters imported from original
parameter file which is subsequently split into model set up and model run time parameters. Data
frame contains probabilities from the scenarios of interest for listing transition rates.}

\item{state_vector}{(class numeric) numeric binary vector of length number of sites containing
information about the state of each site in relation to a condition (e.g. is the site
1 = infected or 0 = susceptible state).}

\item{trans_name}{(class string) string in quotation marks "" stating transition type. Note: this
should correspond to a column name in the input parameter file and is case sensitive. An example
input may be "Site_Recovers".}

\item{site_indices}{(class numeric) vector of 0-based site indices of length 'number of sites'.}

\item{trans_type}{(class numeric) single numeric value indicating the type of transition occurring.}
}
\value{
(class list) of length 5 containing:
\enumerate{
\item (class numeric) numeric vector of transition types (\code{rate_type}).
\item (class integer) integer vector of sites subject to transition (\code{position}).
\item (class numeric) numeric vector of transition rates (\code{rate}).
\item (class logical) logical vector of NA's for infection source (\code{source_site}).
\item (class integer) number of sites in listed transition rates (\code{n_rates}).
}
}
\description{
This function lists transition rates for all site state transition probabilities except for
disease transmission via LFM route of AquaNet-Mod (see details).
}
\details{
Generate a list of transition rate objects using an input vector of sites stating their status
(\code{state_vector}) (for example infection status 1/0, fallow status, latency status) to extract
sites (from \code{site_indices}) where the state in state_vector is 1. For these sites, generate a
vector of the transition rate: \code{1 / run_time_params[[trans_name]]}. The column number of
\code{trans_name} within the \code{run_time_params} depicts the rate type for the site - this is output in
a list alongside the transition rate, transitioning site, rate, source site (NAs), and number of
sites.

Note: in AquaNet-Mod, this function is called multiple times exclusively within the
\code{aquanet::updateRates} function.
}
