% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResultsFullSiteType.R
\name{loadResultsFullSiteType}
\alias{loadResultsFullSiteType}
\title{loadResultsFullSiteType}
\usage{
loadResultsFullSiteType(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded. Assigned
in RunModelCommandLine.R}
}
\value{
(class data.table) sites_summary_type
}
\description{
Loads and wrangles the full data (full_details) outputted from a scenario run.
Returns a data frame with the following information for each timestep:
\enumerate{
\item \code{site_id} (numeric) the site identification number
\item \code{state} (numeric) the state the site is in.
TODO: add file with state definitions and link here
\item \code{sim_no} (numeric) the simulation number
\item \code{t_diff} (numeric) the amount of time the site has spent in the state specified (in days)
\item \code{t} (numeric) the model time
}
}
\details{
Columns 5:20 are logical site type vectors assigned in \code{site_types.csv}-
whether or not a site contains that operation type.

e.g. smallhatch = 1 means the site contains a small hatchery.

e.g. farm_vector = 0 means the site is NOT a farm.

Column 21 is the sum of the site type vector, and is used to diagnosed uncategorised
sites (usually fisheries)
}
