% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRiverDistances.R
\name{loadRiverDistances}
\alias{loadRiverDistances}
\title{loadRiverDistances}
\usage{
loadRiverDistances(
  filepath_river_distances,
  out_createContactProbabilityMatrix
)
}
\arguments{
\item{filepath_river_distances}{(class string) string containing the file path and file name for
.csv containing information about distances between sites via river network proximity and
connectivity generated with GIS tool.}

\item{out_createContactProbabilityMatrix}{(class list) of length 3 containing:
\enumerate{
\item (class integer) number of sites in movements matrix (integer).
\item (class dgCMatrix, Matrix package) movements matrix.
\item (class dgCMatrix, Matrix package) probability of movements matrix with top sites zeroed.
}}
}
\value{
(class data frame) data frame containing the site origin and destination IDs for the river network.
}
\description{
River distances between sites are loaded from .csv whose file name is specified as input \code{filepath_river_distances}.
Origin and destination site IDs are extracted, and the data subsetted to retain only connections
where both origin and destination site are present in the live fish movements contact matrix
(generated with igraph).
}
