% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulations.R
\name{runSimulations}
\alias{runSimulations}
\title{runSimulations}
\usage{
runSimulations(
  n_cores,
  n_sims,
  seed_num,
  clear_results,
  tmax,
  run_time_params,
  non_peak_season_length,
  out_createContactProbabilityMatrix,
  out_createContactProbabilityMatrixTopSitesRemoved,
  out_createWithinCatchmentEdges,
  out_createCatchmentToSiteMatrix,
  out_createRiverDistanceProbabilityMatrix,
  out_createDistanceMatrix,
  farm_vector,
  n_states,
  n_initial_infections,
  type_catchment_controls,
  filepath_results,
  contact_tracing,
  remove_top_sites,
  n_infections_remove_top_sites,
  disease_controls,
  proportion_cullable,
  days_before_catchment_restock,
  river_distances_df,
  site_details,
  stochastic_run,
  seed_farm_choice,
  batch_size = NULL,
  n_sims_per_job = NULL
)
}
\arguments{
\item{n_cores}{(class numeric) the number of cores to use for parallel computing.}

\item{n_sims}{(class numeric) number of simulations.}

\item{batch_size}{(class numeric) number of simulations assigned to each parallel batch job.
This can also be supplied as \code{n_sims_per_job} for compatibility. Defaults to
\code{ceiling(n_sims / n_cores)} when \code{NULL}.}

\item{seed_num}{(class numeric) number used to generate random seed for result replication.}

\item{clear_results}{(class logical) TRUE/FALSE stating whether .RData results should be cleared
from \code{filepath_results} and nested directories. This ensures if you re-run AquaNet-Mod the
results files are all obtained from a single run.}

\item{tmax}{(class numeric) maximum amount of time in days that each simulation should run for.}

\item{run_time_params}{(class data frame) of model run time parameters imported from original
parameter file which is subsequently split into model set up and model run time parameters. Data
frame contains probabilities from the scenarios of interest for listing transition rates.}

\item{non_peak_season_length}{(class string) the amount of time in days that the non peak
transmission season lasts for (e.g. "90", "180").}

\item{out_createContactProbabilityMatrix}{(class list) of length 3 containing (1) number of sites
in (live fish) movements matrix (integer), (2) (live fish) movements matrix (dgCMatrix, Matrix
package), and (3) probability of (live fish) movements matrix (dgTMatrix, Matrix package).}

\item{out_createContactProbabilityMatrixTopSitesRemoved}{(class list) of length 3 containing (1)
number of sites in (live fish) movements matrix (integer), (2) (live fish) movements matrix
(dgCMatrix, Matrix package), and (3) probability of (live fish) movements matrix (dgTMatrix,
Matrix package). This object is created following the removal of the top most connected sites
in the network.}

\item{out_createWithinCatchmentEdges}{(class list) of length 3 containing (1) lgCMatrix (logical
matrix) detailing within catchment connections, (2) edge matrix of vertex IDs within catchments,
and (3) matrix of source site and receiving site within catchment edges.}

\item{out_createCatchmentToSiteMatrix}{(class list) of length 2 containing (1) data frame of site
to catchment information and (2) dgCMatrix sparse matrix containing site to catchment summary.}

\item{out_createRiverDistanceProbabilityMatrix}{(class list) of length 2 containing (1)
distances between sites on a river network (via river connectivity) connections of 0 distances
are removed (data frame) and (2) sparse matrix containing probability of transmission between
sites connected via the river network by river water (dgTMatrix).}

\item{out_createDistanceMatrix}{(class list) of length 3 containing (1) a matrix of site to site
distances (class matrix array), (2) a matrix of distance-based transmission probabilities
(dgTMatrix, Matrix package), and (3) input data frame of site catchment locality
(SpatialPointsDataFrame, sp package).}

\item{farm_vector}{(class numeric) numeric binary vector of length number of sites containing
information on whether each site 1 = is a farm or 0 = is not a farm.}

\item{n_states}{(class numeric) number of different combinations of states possible within the
model.}

\item{n_initial_infections}{(class numeric) number of farms to seed infection at.}

\item{type_catchment_controls}{(class numeric) tnumber selecting catchment level controls to
apply (0 = allows movements within the same catchments, 1 = allows movements within or between
infected catchments, and 2 = allows no movements by any of the sites within an infected
catchment, "None" means there are no catchment level controls).}

\item{filepath_results}{(class character) character vector containing paths to various results
directories for model run created in \code{aquanet-mod/code/RunModelCommandLine.R}.}

\item{contact_tracing}{(class logical) vector of length 1 indicating whether or not contact
tracing should take place.}

\item{remove_top_sites}{(class logical) vector of length 1 indicating whether or not the removal
of the most connected sites in the network should take place.}

\item{n_infections_remove_top_sites}{(class numeric) vector of length 1. After the cumulative
number of infected sites exceeds this number, switch to using the top sites removed contact
probability matrix.}

\item{disease_controls}{(class logical) vector of length 1 indicating whether or not
any disease control measures should take place.}

\item{proportion_cullable}{(class numeric) proportion of fisheries able to cull site.}

\item{days_before_catchment_restock}{(class numeric) number of days that all sites within a
catchment need to be in the post-fallow state before restocking.}

\item{river_distances_df}{(class data frame) a data frame of distances between sites
along the river network. Created using the GIS tool.}

\item{site_details}{(class data frame) a data frame of site and model IDs, locations and whether
or not the site is tidal. Created using \code{aquanet::mergeGraphMetaWithCatchmentLocation}.}

\item{stochastic_run}{(class logical) vector of length 1 indicating whether
to randomly seed the starting farm (TRUE) or manually select it.}

\item{seed_farm_choice}{(class character) character vector of length 1 containing
the manual selection for seed farm. Will be used if \code{stochastic_run} is FALSE.}
}
\value{
(class list) of length 2 containing:
\enumerate{
\item (class numeric) the number of cores used for the run.
\item the output of the foreach loop running the \code{aquanet::simulationCode()} function.
}
}
\description{
This function runs the code to simulate disease transmission throughout a network of aquaculture
sites through time using parallel computing (see details).
}
\details{
If \code{clear_results} = TRUE, the runSimulations function first clears all .RData files within the
\code{filepath_results} and daughter directories.

The number of simulations per job is calculated and a set of R copies running in parallel,
communicating over sockets is generated.

The \code{aquanet::simulationCode()} function is then called in parallel to run the AquaNet-Mod
simulation with defined inputs. One complete, the parellel R copies are stopped.
}
