% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateCount.R
\name{stateCount}
\alias{stateCount}
\title{stateCount}
\usage{
stateCount(no_per_state_wide)
}
\arguments{
\item{no_per_state_wide}{(class data.frame) returned from \code{loadResultsSummary}}
}
\value{
(class data.frame) sum_per_state
}
\description{
Returns a tally of the number of times a state appears in the summary table.
}
\details{
Useful for checking you have the correct outputs from a scenario run.
}
