% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePerStage.R
\name{timePerStage}
\alias{timePerStage}
\title{timePerStage}
\usage{
timePerStage(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded. Assigned
in RunModelCommandLine.R}
}
\value{
(data.table) sims_site_time_summary
}
\description{
Summarises the output of \code{loadResultsFullSiteType} so that each row is the
total amount of time (\code{t_total}) a site (\code{site_id}) spends in a state (\code{state})
per simulation (\code{sim_no}). Also included are the site categories.
}
\details{
Used for economic analyses
}
