% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlots.R
\name{timePlot}
\alias{timePlot}
\title{timePlots}
\usage{
timePlot(scenario_name, scenario_summary, y_max)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded.}

\item{scenario_summary}{(class string) a file path to the scenario summary produced by cumulativeTimeSummary}

\item{y_max}{(class numeric) maximum number of infected sites}
}
\value{
outputPlot
}
\description{
Time plots created displaying cumulative number of infected sites on the y axis and time in days on the x axis.
}
