% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissionRouteProportions.R
\name{transmissionRouteProportions}
\alias{transmissionRouteProportions}
\title{transmissionRouteProportions}
\usage{
transmissionRouteProportions(scenario_name)
}
\arguments{
\item{scenario_name}{The name of the scenario to be checked.}
}
\value{
(class data.table) A summary of the infections per scenario, in terms of count and percentage
}
\description{
Get the proportions of infections via each possible mechanism for each scenario.
Possible mechanisms are:
\enumerate{
\item LFM = live fish movement
\item RB = river-based transmission
\item SDM = short-distance mechanical
\item DIM = distance-independent mechanical (random spillover)
}
}
