% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstreamSiteRelease.R
\name{upstreamSiteRelease}
\alias{upstreamSiteRelease}
\title{upstreamSiteRelease}
\usage{
upstreamSiteRelease(downstream_sites, sites_unique, sites_controlled)
}
\arguments{
\item{downstream_sites}{(class vector) a vector of downstream sites, created
by the \code{createRiverDistanceProbabilityMatrix} function.}

\item{sites_unique}{(class vector) a vector of unique site IDs.}

\item{sites_controlled}{(class vector) a vector of sites currently under catchment-level
controls (0 = FALSE, 1 = TRUE).}
}
\value{
(class vector) a vector of sites indicating whether or not they are under
catchment-level controls (0 = FALSE, 1 = TRUE).
}
\description{
This function is used when the catchment-level controls include the release of
sites upstream of infected sites.
}
\details{
First, a vector of upstream sites is identified, where 0 = downstream and 1 =
upstream.
Second, upstream sites are removed from \code{downstream_sites_controlled}, which
gives the sites under catchment-level controls.
The output is then used to adjust the contact probability matrix.
}
