% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateCosts.R
\docType{data}
\name{value}
\alias{value}
\title{stateCosts}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
value
}
\arguments{
\item{data}{(class data.frame) output of timePerStage. Contains the site ID,
state, simulation number, total time, and the logical vector of site types.}

\item{state}{(class character) describes the state for which the costs will be
calculated. Accepted inputs are: \code{"fallow"} in a fallow state (infected or uninfected),
\code{"no_manage"} infected in an unmanaged state, \code{"contact_trace"} in a contact tracing state
(infected or uninfected), or \code{"catchment_control"} uninfected site under catchment-level
controls.}

\item{site_types}{(class character) a vector of possible site types.}
}
\value{
(class list) of length 2 containing:
\enumerate{
\item (class data.table) a full breakdown of daily costs by site type and simulation.
\item (class data.table) a summary data frame containing the simulation number and the cost
incurred by sites being in that state.
}
}
\description{
Calculates the total duration costs for each simulation across fallow, contact tracing,
catchment control and unmanaged disease scenarios.
}
\keyword{datasets}
