% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexBetweenness.R
\name{vertexBetweenness}
\alias{vertexBetweenness}
\title{Vertex Betweenness summary statistics
Vertex Betweenness summary statistics. Indicating the measure of centrality in a graph based on shortest paths.}
\usage{
vertexBetweenness(contact_network)
}
\arguments{
\item{contact_network}{network simulates and tracks disease transmission from infected to susceptible contacts}
}
\value{
network_betweenness
}
\description{
Vertex Betweenness summary statistics
Vertex Betweenness summary statistics. Indicating the measure of centrality in a graph based on shortest paths.
}
