% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexCloseness.R
\name{vertexCloseness}
\alias{vertexCloseness}
\title{Vertex Closeness summary
Summary statistics of vertex closeness centrality for the contact network.The closeness centrality of a vertex is the inverse of the sum of the distances from a vertex (point in a graph) to all the other vertices.}
\usage{
vertexCloseness(contact_network)
}
\arguments{
\item{contact_network}{network simulates and tracks disease transmission from infected to susceptible contacts}
}
\value{
network_closeness
}
\description{
Vertex Closeness summary
Summary statistics of vertex closeness centrality for the contact network.The closeness centrality of a vertex is the inverse of the sum of the distances from a vertex (point in a graph) to all the other vertices.
}
