% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSiteTypeVector.R
\name{createSiteTypeVector}
\alias{createSiteTypeVector}
\title{createSiteTypeVector}
\usage{
createSiteTypeVector(graph_name, site_type)
}
\arguments{
\item{graph_name}{(class igraph) the site contact network generated by the script
\verb{code/03_CreateContactNetwork.R}.}

\item{site_type}{(class character) the site type to be assigned.}
}
\value{
(class vector) logical vector of length 'number of sites' stating whether each site is
of type \code{site_type}.
}
\description{
Function creates a logical vector (0 = FALSE, 1 = TRUE) of whether a site belongs to a given
type needed for \code{code/Economics.R}.
}
