% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cullCost.R
\name{cullCost}
\alias{cullCost}
\title{cullCost}
\usage{
cullCost(farm_data, cull_costs, site_types)
}
\arguments{
\item{farm_data}{(class data.frame) the \code{time_summary} data frame created from the
\code{aquanet::importAndCondense()} function, for farms only.}

\item{cull_costs}{(class data.frame) of the costs to both the site and the competent authority
for each site type.}

\item{site_types}{(class character)  a vector of possible site types.}

\item{non_farm_data}{(class data.frame) the \code{time_summary} data frame created from the
\code{aquanet::importAndCondense} function, for non-farms (fisheries) only.}
}
\value{
(class data.table) the cost to cull across each simulation.
}
\description{
Calculates a one-off cost to both the site and the competent authority in the case of a site
needing to be culled. Returns a data frame the cost to cull across each simulation.
}
