% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAndCondense.R
\name{importAndCondense}
\alias{importAndCondense}
\title{importAndCondense}
\usage{
importAndCondense(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded. Assigned
in the params.yaml}
}
\value{
(class data.table) sites_summary_type
}
\description{
Loads and condenses the full data (full_details) outputted from a scenario run.
Returns a data frame with the following information for each site in each simulation:
\enumerate{
\item \code{site_id} (numeric) the site identification number
\item \code{state} (numeric) the state the site is in.
TODO: add file with state definitions and link here
\item \code{sim_no} (numeric) the simulation number
\item \code{t_total} (numeric) the total amount of time the site has spent in the state specified (in days)
\item \code{trans_type} (numeric) the transition type which is occuring at that timestep
}
}
\details{
Columns 6:21 are logical site type vectors assigned in \code{site_types.csv}-
whether or not a site contains that operation type.

e.g. smallhatch = 1 means the site contains a small hatchery.

e.g. farm_vector = 0 means the site is NOT a farm.

Column 21 is the sum of the site type vector, and is used to diagnosed uncategorised
sites (usually fisheries)
}
