% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importsExports.R
\name{importsExports}
\alias{importsExports}
\title{importsExports
Summarise the  movement type of nodes (susceptible contacts)
as either imports (the number of incoming connections),
or outputs (the number of outgoing connections ).
Then identify the node with the largest number of outputs.}
\usage{
importsExports(contact_network, imports_exports)
}
\arguments{
\item{contact_network}{(class igraph)  Graph of connections/movements between sites produced with iGraph
in '03_CreateContactNetwork.R' of AquaNet-mod.}

\item{imports_exports}{string input of either 'imports' (in-degree movements) or 'exports'(out-degree movements)}
}
\value{
movements
}
\description{
importsExports
Summarise the  movement type of nodes (susceptible contacts)
as either imports (the number of incoming connections),
or outputs (the number of outgoing connections ).
Then identify the node with the largest number of outputs.
}
