% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTransitionRates.R
\name{listTransitionRates}
\alias{listTransitionRates}
\title{listTransitionRates}
\usage{
listTransitionRates(
  run_time_params,
  state_vector,
  trans_type,
  site_indices,
  infection_state
)
}
\arguments{
\item{run_time_params}{(class data frame) of model run time parameters imported from original
parameter file which is subsequently split into model set up and model run time parameters. Data
frame contains probabilities from the scenarios of interest for listing transition rates.}

\item{state_vector}{(class numeric) numeric binary vector of length number of sites containing
information on whether each site is in an 1 = infected or 0 = susceptible state.}

\item{trans_type}{(class string) string in quotation marks "" stating transition type. Note: this
should correspond to a column name in the input parameter file and is case sensitive. An example
input may be "Site_Recovers".}

\item{site_indices}{(class numeric) vector of 0-based site indices of length 'number of sites'.}

\item{infection_state}{(class numeric) infection status of sites whose transition rates should be
determined to subset either 1 = infected or 0 = susceptible.}
}
\value{
(class list) of length 5 containing:
\enumerate{
\item (class numeric) numeric vector of transition types (\code{rate_type}).
\item (class integer) integer vector of sites subject to transition (\code{position}).
\item (class numeric) numeric vector of transition rates (\code{rate}).
\item (class logical) logical vector of NA's for infection source (\code{source_site}).
\item (class integer) number of sites in listed transition rates (\code{n_rates}).
}
}
\description{
Generate a list of transition rate objects using an input vector of sites stating their infection
status (\code{state_vector}) to extract sites (from \code{site_indices}) with a specified infection status
(susceptible/infected) \code{infection_state}). For these sites, generate a vector of the transition
rate: \code{1 / run_time_params[[trans_type]]}. The column number of \code{trans_type} within the
\code{run_time_params} depicts the rate type for the site - this is output in a list alongside the
transition rate, transitioning site, rate, source site (NAs), and number of sites.
}
\details{
Note: in aquanet-mod, this function is called multiple times exclusively within the \code{update_rate}
function.
}
