% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResultsSummary.R
\name{loadResultsSummary}
\alias{loadResultsSummary}
\title{loadResultsSummary}
\usage{
loadResultsSummary(scenario_name)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded. Assigned
in RunModelCommandLine.R}
}
\value{
(class data.frame) no_per_state_wide
}
\description{
Loads and wrangles the batch (summary) data outputted from a scenario run.
Returns a data frame with the following information for each timestep:
\enumerate{
\item \code{batch_no} the batch number
\item \code{k} the run number within the batch
\item \code{t} the model time
\item \code{t_diff} the difference between the current time and the time at the previous timestep
(the duration of the previous timestep)
\item \code{sim_no} the simulation number
\item \code{rate_type} the transition type
\item \code{no_controlled_catchments} the number of catchments under controls in that timestep
\item \code{cumulative_no_infected_sites} the cumulative number of infected sites in the simulation
}
}
