% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationCode.R
\name{simulationCode}
\alias{simulationCode}
\title{simulationCode}
\usage{
simulationCode(
  graph.contactp.objects,
  runs,
  tmax,
  batchNo,
  ListRunTimeParameters,
  graph.withinCatchmentEdges.objects,
  graph.catchment2Site.objects,
  graph.riverDistance.objects,
  graph.estimateSiteDistances.objects,
  farm_vector,
  associatedSiteControlType,
  locationSaveResults,
  initialNoInfections
)
}
\arguments{
\item{graph.contactp.objects}{TODO}

\item{runs}{TODO}

\item{tmax}{TODO}

\item{batchNo}{TODO}

\item{ListRunTimeParameters}{TODO}

\item{graph.withinCatchmentEdges.objects}{TODO}

\item{graph.catchment2Site.objects}{TODO}

\item{graph.riverDistance.objects}{TODO}

\item{graph.estimateSiteDistances.objects}{TODO}

\item{farm_vector}{TODO}

\item{associatedSiteControlType}{TODO}

\item{locationSaveResults}{TODO}

\item{initialNoInfections}{TODO}
}
\value{

}
\description{
simulationCode
}
