% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationCode.R
\name{simulationCode}
\alias{simulationCode}
\title{simulationCode}
\usage{
simulationCode(
  createContactProbabilityMatrix_out,
  runs,
  tmax,
  batch_num,
  run_time_params,
  non_peak_season_length,
  createWithinCatchmentEdges_out,
  createCatchmentToSiteMatrix_out,
  createRiverDistanceProbabilityMatrix_out_list,
  createDistanceMatrix_out,
  farm_vector,
  type_catchment_controls,
  filepath_results
)
}
\arguments{
\item{createContactProbabilityMatrix_out}{TODO (was graph.contactp.objects)}

\item{runs}{TODO}

\item{tmax}{TODO}

\item{batch_num}{TODO (was batchNo)}

\item{run_time_params}{TODO}

\item{non_peak_season_length}{TODO}

\item{createWithinCatchmentEdges_out}{TODO (was graph.withinCatchmentEdges.objects)}

\item{createCatchmentToSiteMatrix_out}{TODO}

\item{createRiverDistanceProbabilityMatrix_out_list}{TODO}

\item{createDistanceMatrix_out}{TODO (was graph.estimateSiteDistances.objects)}

\item{farm_vector}{TODO}

\item{type_catchment_controls}{TODO}

\item{filepath_results}{TODO}
}
\value{

}
\description{
simulationCode
}
