% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlots.R
\name{timePlot}
\alias{timePlot}
\title{timePlots}
\usage{
timePlot(scenario_name, y_max)
}
\arguments{
\item{scenario_name}{(class string) the name of the scenario being loaded.}

\item{y_max}{maximum number of infected sites}
}
\value{
outputPlot
}
\description{
Time plots created displaying cumulative number of infected sites on the y axis and time in days on the x axis.
}
