% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstreamSiteRelease.R
\name{upstreamSiteRelease}
\alias{upstreamSiteRelease}
\title{upstreamSiteRelease}
\usage{
upstreamSiteRelease(
  river_distances_df,
  move_restricted_sites,
  sites_tidal,
  sites_unique,
  sites_controlled
)
}
\arguments{
\item{river_distances_df}{(class data frame) a data frame of distances between sites
along the river network. Created using the GIS tool.}

\item{move_restricted_sites}{(class logical) logical vector of length 'number of sites' that
states whether movements at this site are currently restricted (TRUE) or unrestricted (FALSE).}

\item{sites_tidal}{(class data frame) logical data frame of length 'number of sites' that
states whether this site is located within a tidal boundary (TRUE) or not (FALSE).
(Note: created in the \code{aquanet::updateRates} function of aquanet-mod).}

\item{sites_unique}{(class vector) a vector of unique site IDs.}

\item{sites_controlled}{(class vector) a vector of sites currently under catchment-level
controls (0 = FALSE, 1 = TRUE).}
}
\value{
(class vector) a vector of sites indicating whether or not they are under
catchment-level controls (0 = FALSE, 1 = TRUE).
}
\description{
This function is used when the catchment-level controls include the release of
sites upstream of infected sites. It identifies "safe sites" upstream of the infected
site so they may be removed from controls.
}
