% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexBetweenness.R
\name{vertexBetweenness}
\alias{vertexBetweenness}
\title{vertexBetweenness
Summary of vertex betweenness of contact network, which ranks which nodes (susceptible contacts) are significant,
based on the fraction of shortest paths that pass through each node.}
\usage{
vertexBetweenness(contact_network)
}
\arguments{
\item{contact_network}{(class igraph)  Graph of connections/movements between sites produced with iGraph
in '03_CreateContactNetwork.R' of AquaNet-mod.}
}
\value{
network_betweenness
}
\description{
vertexBetweenness
Summary of vertex betweenness of contact network, which ranks which nodes (susceptible contacts) are significant,
based on the fraction of shortest paths that pass through each node.
}
