% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexCloseness.R
\name{vertexCloseness}
\alias{vertexCloseness}
\title{Vertex Closeness
Summary of vertex closeness of contact network, which ranks which nodes (susceptible contacts) are significant,
based on the length of the shortest paths between a node to all other nodes in the network.}
\usage{
vertexCloseness(contact_network)
}
\arguments{
\item{contact_network}{(class igraph)  Graph of connections/movements between sites produced with iGraph
in '03_CreateContactNetwork.R' of AquaNet-mod.}
}
\value{
network_closeness
}
\description{
Vertex Closeness
Summary of vertex closeness of contact network, which ranks which nodes (susceptible contacts) are significant,
based on the length of the shortest paths between a node to all other nodes in the network.
}
